fix(blsct): four small correctness fixes in the proof helpers
Each of these is independent; none changes a serialized format or a
consensus-visible value.

bulletproofs_plus::RangeProof::operator== ignored tau_x, which the same
class serializes and the non-plus variant compares, so two proofs
differing only in that scalar compared equal.

Elements::ConfirmIndexInsideRange took a uint32_t while its only callers,
the operator[] overloads, take a size_t. An index above 2^32 was narrowed
before the bounds check and then used to index the vector -- the test for
it segfaults without this change rather than merely missing a throw.

GetFirstPowerOf2GreaterOrEqTo documents that it throws when no such power
is representable. It did not: `i *= 2` wrapped to zero and the loop then
ran forever. Make the code match the contract; every input that worked
before returns the same value.

HDChain::operator== compared blinding_id, which SERIALIZE_METHODS never
writes, so a chain never equalled its own reloaded self. Drop the field
from the comparison rather than change the record format.

+++ src/blsct/arith/elements.h
@@ -39,7 +39,7 @@ public:
     bool Find(const T& x) const;
     std::vector<uint8_t> GetVch() const;
 
-    void ConfirmIndexInsideRange(const uint32_t& index) const;
+    void ConfirmIndexInsideRange(const size_t& index) const;
     void ConfirmSizesMatch(const size_t& other_size) const;
 
     /**
